requirements.txt
# Core dependencies
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Email processing
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pytz==2023.3

# Todoist API
todoist-api-python==2.1.3

# Utils
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17

# Logging
colorlog==6.8.0 